backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0
opencv-python>=4.8.0
openpyxl>=3.1.0
python-multipart>=0.0.6
setuptools>=68.0.0
scipy>=1.10.0
